GeoWeb Service For Census Data Management
The mode in which census data is presented is a key role in acquiring census data. In this project a web service is created in django where census data can be acquired interactively. Choropleth maps are used to show census data with user interactions. Data is also presented in tabular form.